refactor: split oversized update files
